safari-600.7-branch/LayoutTests/fast/regions/cssom/webkit-named-flow-get-regions-expected.txt

Tests WebKitNamedFlow.getRegions() and the DOM order of the list

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Flow content
Test if the regions are in DOM order
PASS getRegionAt(0).id is "region_1"
PASS getRegionAt(1).id is "region_2"
Empty content
PASS successfullyParsed is true

TEST COMPLETE
 
